experiments/experiments_requirements.txt
# Skip steps 1 and 2 if using Docker container to run the experiments.
# 1. Run manually one of the following:
#    If using conda environment: conda install -c "nvidia/label/cuda-11.8.0" cuda-toolkit
#    Otherwise: sudo apt-get install cuda
# 2. Follow https://docs.nvidia.com/cuda/cuda-installation-guide-linux/index.html#environment-setup to set the
#    environment variables PATH and LD_LIBRARY_PATH. To find the lib/bin folders
#    in case of conda environment use: find ~ -name 'libcusolver.so.11'
#    and in case of a global installation with apt-get: find /usr/ -name 'libcusolver.so.11'
#    and use the containing folder.

# Installs JAX's wheel compatible with CUDA 11 and cuDNN 8.2 or newer
#-f https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
jax[cuda11_cudnn86]==0.4.2
optax==0.1.4
functorch==0.2.1
torchopt==0.5.0.post4
torch=1.12.1

# Packages for the experiments
tensorflow
tensorflow-datasets
dm-haiku
torchvision
matplotlib
flax
pandas
jax-resnet
numpy
tqdm
scipy
soundfile
resampy
kfac-jax
jaxopt